pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-google-genai>=2.1.3",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.3.31",
    "mcp[cli]>=1.6.0",
    "numexpr>=2.10.2",
    "python-dotenv>=1.1.0",
]
